macro_tools : fixing mess
