+++ module/core/macro_tools/tests/inc/basic_test.rs
@@ -343,64 +343,6 @@ TokenStream [
 
   }
 
-  //
-
-  #[ test ]
-  fn attr_pair_single_basic() -> Result< () >
-  {
-    use syn::spanned::Spanned;
-
-    // test.case( "basic" );
-    let input = qt!
-    {
-      #[ derive( Former ) ]
-      pub struct Struct1
-      {
-        #[former( default = 31 ) ]
-        pub int_1 : i32,
-      }
-    };
-
-    let ast = match syn::parse2::< syn::DeriveInput >( input )
-    {
-      Ok( syntax_tree ) => syntax_tree,
-      Err( err ) => return Err( err ),
-    };
-
-    let fields = match ast.data
-    {
-      syn::Data::Struct( ref data_struct ) => match data_struct.fields
-      {
-        syn::Fields::Named( ref fields_named ) =>
-        {
-          &fields_named.named
-        },
-        _ => return Err( syn::Error::new( ast.span(), "Unknown format of data, expected syn::Fields::Named( ref fields_named )" ) ),
-      },
-      _ => return Err( syn::Error::new( ast.span(), "Unknown format of data, expected syn::Data::Struct( ref data_struct )" ) ),
-    };
-
-    let attr = fields.first().ok_or_else( || err( "No field" ) )?.attrs.first().ok_or_else( || err( "No attr" ) )?;
-
-    let exp = Equation
-    {
-      left : parse_quote!{ default },
-      op : parse_quote!{ = },
-      right : parse_quote!{ 31 },
-    };
-    let got = the_module::equation( &attr )?;
-    a_id!( got.left, exp.left );
-    a_id!( format!( "{:?}", got ), format!( "{:?}", exp ) );
-    // a_id!( got.right, exp.right );
-
-    return Ok( () );
-
-    fn err( src : &str ) -> syn::Error
-    {
-      syn::Error::new( proc_macro2::Span::call_site(), src )
-    }
-  }
-
 }
 
 //
@@ -413,5 +355,4 @@ tests_index!
   type_optional_container_kind_basic,
   type_rightmost_basic,
   type_parameters_basic,
-  attr_pair_single_basic,
 }
