+++ module/core/macro_tools/tests/inc/mod.rs
@@ -12,20 +12,22 @@ mod if_enabled
 
   use super::*;
   use the_module::exposed::*;
+  // xxx : use protected
 
-  mod attr;
-  mod basic;
-  mod derive;
-  mod drop;
-  mod generic_args;
-  mod generic_params;
-  mod item;
-  mod item_struct;
-  mod phantom;
-  mod quantifier;
-  mod struct_like;
-  mod syntax;
-  mod tokens;
-  mod typ;
+  mod attr_test;
+  mod basic_test;
+  mod derive_test;
+  mod drop_test;
+  mod equation_test;
+  mod generic_args_test;
+  mod generic_params_test;
+  mod item_test;
+  mod item_struct_test;
+  mod phantom_test;
+  mod quantifier_test;
+  mod struct_like_test;
+  mod syntax_test;
+  mod tokens_test;
+  mod typ_test;
 
 }
